Added a Cleaner and Various Options
- a cleaner daemon runs to clean the orphaned files (docker.env has the configuration parameter API_CLEANER_FILE_LIFETIME to establish when a token expires);
- the option to choose the hashing algorithm (also in docker.env has the configuration parameter API_IMAGE_HASHER and you can choose between SHA256 and MD5);

+++ api.py
@@ -3,7 +3,7 @@ import os
 from io import BytesIO
 from flask import Flask, request, Response, send_file
 from pathlib import Path
-from hashlib import sha256
+from hashlib import md5, sha256
 import base64
 import json
 
@@ -22,7 +22,10 @@ def put_image():
     image_type = image_file.mimetype
     image_data = image_file.read()
 
-    image_hash = sha256()
+    image_hash = ({
+        'MD5': md5,
+        'SHA256': sha256
+    }.get(os.environ.get('API_IMAGE_HASHER', 'SHA256').upper()) or sha256)()
     image_hash.update(image_data)
     image_token = image_hash.hexdigest()
 
@@ -53,6 +56,9 @@ def get_image():
     STAGED_PATH = os.environ.get("STAGED_PATH", "/tmp/ai/staged")
     SOURCE_PATH = os.environ.get("SOURCE_PATH", "/tmp/ai/source")
     PREPARED_PATH = os.environ.get("PREPARED_PATH", "/tmp/ai/prepared")
+    FILE_LIFETIME = os.environ.get('API_CLEANER_FILE_LIFETIME', '1800.0')
+
+    lifetime = float(FILE_LIFETIME)
 
     if image_type == 'image/png':
         staged_file = Path(STAGED_PATH) / (image_token + '.png')
@@ -67,9 +73,9 @@ def get_image():
     else:
         return Response(json.dumps({'error': 'unknown image format'}), status = 400, mimetype='application/json')
 
-    if prepared_file.is_file():
-        with prepared_file.open('rb') as fp:
-            image_data = fp.read()
+    token_expired = staged_file.stat().st_mtime + lifetime < time.time()
+
+    if prepared_file.is_file() and not token_expired:
         if staged_file.is_file():
             staged_file.unlink()
         if source_file.is_file():
@@ -80,11 +86,23 @@ def get_image():
 
         return send_file(BytesIO(image_data), mimetype = 'image/png', as_attachment = True, download_name = prepared_file.name)
 
-    if staged_file.is_file():
+    if staged_file.is_file() and not token_expired:
         return Response(json.dumps({'wait': 'true', 'status': 'not queued'}), status = 200, mimetype = 'application/json')
 
-    if source_file.is_file():
+    if source_file.is_file() and not token_expired:
         return Response(json.dumps({'wait': 'true', 'status': 'processing'}), status = 200, mimetype = 'application/json')
+
+    if token_expired:
+        if staged_file.is_file():
+            staged_file.unlink()
+        if source_file.is_file():
+            source_file.unlink()
+        if json_file.is_file():
+            json_file.unlink()
+        if prepared_file.is_file():
+            prepared_file.unlink()
+
+        return Response(json.dumps({'error': 'token expired'}), status = 400, mimetype = 'application/json')
 
     return Response(json.dumps({'error': 'unknown token'}), status = 400, mimetype = 'application/json')
 
@@ -96,6 +114,9 @@ def get_json():
     STAGED_PATH = os.environ.get("STAGED_PATH", "/tmp/ai/staged")
     SOURCE_PATH = os.environ.get("SOURCE_PATH", "/tmp/ai/source")
     PREPARED_PATH = os.environ.get("PREPARED_PATH", "/tmp/ai/prepared")
+    FILE_LIFETIME = os.environ.get('API_CLEANER_FILE_LIFETIME', '1800.0')
+
+    lifetime = float(FILE_LIFETIME)
 
     if image_type == 'image/png':
         staged_file = Path(STAGED_PATH) / (image_token + '.png')
@@ -109,8 +130,10 @@ def get_json():
         json_file = Path(PREPARED_PATH) / (image_token + '.json')
     else:
         return Response(json.dumps({'error': 'unknown image format'}), status = 400, mimetype='application/json')
+    
+    token_expired = staged_file.stat().st_mtime + lifetime < time.time()
 
-    if json_file.is_file():
+    if json_file.is_file() and not token_expired:
         with json_file.open('rb') as fp:
             json_data = json.load(fp)
         if staged_file.is_file():
@@ -128,11 +151,23 @@ def get_json():
 
         return Response(json.dumps(json_data), status = 200, mimetype='application/json')
 
-    if staged_file.is_file():
+    if staged_file.is_file() and not token_expired:
         return Response(json.dumps({'wait': 'true', 'status': 'not queued'}), status = 200, mimetype = 'application/json')
 
-    if source_file.is_file():
+    if source_file.is_file() and not token_expired:
         return Response(json.dumps({'wait': 'true', 'status': 'processing'}), status = 200, mimetype = 'application/json')
+
+    if token_expired:
+        if staged_file.is_file():
+            staged_file.unlink()
+        if source_file.is_file():
+            source_file.unlink()
+        if json_file.is_file():
+            json_file.unlink()
+        if prepared_file.is_file():
+            prepared_file.unlink()
+
+        return Response(json.dumps({'error': 'token expired'}), status = 400, mimetype = 'application/json')
 
     return Response(json.dumps({'error': 'unknown token'}), status = 400, mimetype = 'application/json')
